refactor(scripts): extract shared GA4 service-account auth helper
getServiceAccountToken/fetchRetry/sleep/DEFAULT_GA4_PROPERTY_ID were
copy-pasted across analytics-report.mjs and the 2 new user-value-metric
scripts. Extracted to scripts/lib/ga4-service-account.mjs per AGENTS.md
sibling-pattern-fix discipline. Verified byte-identical live behavior
pre/post-refactor and 108/108 vitest green.
